Test: guest_cannot_create_and_delete_pizzas

+++ app/Http/Controllers/PizzaController.php
@@ -3,12 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\Models\Pizza;
+use Auth;
 use Illuminate\Http\Request;
 
 class PizzaController extends Controller
 {
     public function store(Request $request)
     {
+        if (Auth::guest()) {
+            return response("Unauthorized", 401);
+        }
         $validatedData = $request->validate([
             'name' => 'required|unique:App\Models\Pizza,name'
         ]);
@@ -17,6 +21,9 @@ class PizzaController extends Controller
 
     public function destroy(Request $request, Pizza $pizza)
     {
+        if (Auth::guest()) {
+            return response("Unauthorized", 401);
+        }
         $pizza->delete();
     }
 }
